+++ src/main/res/xml/pref_all.xml
@@ -223,9 +223,9 @@
         <Preference android:key="aclupdate" android:title="@string/aclupdate"
             android:summary="@string/aclupdate_summary"/>
         <Preference android:key="frontproxy" android:title="@string/frontproxy_set"/>
-        <Preference android:key="about" android:title="@string/about"/>
+<!--        <Preference android:key="about" android:title="@string/about"/>-->
 
-        <Preference android:key="logcat" android:title="Logcat"/>
+<!--        <Preference android:key="logcat" android:title="Logcat"/>-->
 
     </PreferenceCategory>
