+++ src/main/scala/com/github/shadowsocks/fragments/SettingFragment.scala
@@ -1,20 +1,32 @@
 package com.github.shadowsocks.fragments
 
-import android.content.SharedPreferences
+import java.io.{BufferedReader, InputStreamReader}
+import java.util.Locale
+
+import android.content.{Intent, SharedPreferences}
 import android.content.SharedPreferences.OnSharedPreferenceChangeListener
+import android.net.Uri
 import android.os.Bundle
-import android.preference.{CheckBoxPreference, ListPreference, PreferenceFragment, PreferenceManager}
+import android.preference.{CheckBoxPreference, ListPreference, Preference, PreferenceFragment, PreferenceManager}
+import android.support.v7.app.AlertDialog
 import android.util.Log
-import com.github.shadowsocks.{R, SettingActivity, Shadowsocks}
+import android.webkit.{WebView, WebViewClient}
+import android.widget.EditText
+import com.github.shadowsocks.{BuildConfig, R, SettingActivity, Shadowsocks}
 import com.github.shadowsocks.utils.Key
 import com.github.shadowsocks.ShadowsocksApplication.app
+import com.github.shadowsocks.ShadowsocksSettings.TAG
 import com.github.shadowsocks.preferences.DropDownPreference
+import tun2socks.Tun2socks
+
+import scala.collection.mutable
 
 
 class SettingFragment extends PreferenceFragment with OnSharedPreferenceChangeListener {
   lazy val sortMethod = findPreference(Key.SORT_METHOD).asInstanceOf[DropDownPreference]
   lazy val hideServer = findPreference(Key.HIDE_SERVER).asInstanceOf[CheckBoxPreference]
   lazy val autoUpdate = findPreference(Key.AUTO_UPDATE_SUBSCRIPTION).asInstanceOf[CheckBoxPreference]
+  lazy val aboutPref = findPreference("about")
   lazy val prefs = PreferenceManager.getDefaultSharedPreferences(activity)
   private def activity = getActivity.asInstanceOf[SettingActivity]
 
@@ -36,6 +48,64 @@ class SettingFragment extends PreferenceFragment with OnSharedPreferenceChangeLi
       val autoUpdateValue = value.asInstanceOf[Boolean]
       prefs.edit().putBoolean(Key.AUTO_UPDATE_SUBSCRIPTION, autoUpdateValue).apply()
       prefs.edit().putInt(Key.ssrsub_autoupdate, if (autoUpdateValue) 1 else 0).apply()
+      true
+    })
+
+    aboutPref.setSummary(s"v${BuildConfig.VERSION_NAME}")
+    aboutPref.setOnPreferenceClickListener(_ => {
+      val web = new WebView(activity)
+      web.loadUrl("file:///android_asset/pages/about.html")
+      web.setWebViewClient(new WebViewClient() {
+        override def shouldOverrideUrlLoading(view: WebView, url: String): Boolean = {
+          try {
+            startActivity(new Intent(Intent.ACTION_VIEW, Uri.parse(url)))
+          } catch {
+            case _: android.content.ActivityNotFoundException => // Ignore
+          }
+          true
+        }
+      })
+
+      new AlertDialog.Builder(activity)
+        .setTitle(getString(R.string.about_title).formatLocal(Locale.ENGLISH, BuildConfig.VERSION_NAME))
+        .setNegativeButton(getString(android.R.string.ok), null)
+        .setView(web)
+        .create()
+        .show()
+      true
+    })
+
+    findPreference("logcat").setOnPreferenceClickListener(_ => {
+      val et_logcat = new EditText(activity)
+
+      try {
+        val lst = new mutable.LinkedHashSet[String]()
+        lst.add("logcat")
+        lst.add("-d")
+        lst.add("-v")
+        lst.add("time")
+        lst.add("com.xxf098.ssrray")
+        val logcat = Runtime.getRuntime.exec(lst.toArray)
+        val br = new BufferedReader(new InputStreamReader(logcat.getInputStream()))
+        var line = ""
+        line = br.readLine()
+        while (line != null) {
+          et_logcat.append(line)
+          et_logcat.append("\n")
+          line = br.readLine()
+        }
+        br.close()
+      } catch {
+        case e: Exception =>  // unknown failures, probably shouldn't retry
+          e.printStackTrace()
+      }
+
+      new AlertDialog.Builder(activity)
+        .setTitle("Logcat")
+        .setNegativeButton(getString(android.R.string.ok), null)
+        .setView(et_logcat)
+        .create()
+        .show()
       true
     })
